Salient v11.0.2

+++ nectar/helpers/nav-menus.php
@@ -14,6 +14,11 @@ if ( ! defined( 'ABSPATH' ) ) {
 
 
 
+/**
+ * Register theme menu locations.
+ *
+ * @since 1.0
+ */
 if ( function_exists( 'register_nav_menus' ) ) {
 
 	function nectar_add_theme_menu_locations() {
@@ -24,11 +29,13 @@ if ( function_exists( 'register_nav_menus' ) ) {
 		$usingPRCompatLayout           = false;
 		$usingTopLeftRightCompatLayout = false;
 
-		if ( ! empty( $nectar_options['header_format'] ) && $nectar_options['header_format'] == 'menu-left-aligned' || $nectar_options['header_format'] == 'centered-menu' ) {
+		if ( ! empty( $nectar_options['header_format'] ) && $nectar_options['header_format'] === 'menu-left-aligned' 
+		|| $nectar_options['header_format'] === 'centered-menu' 
+		|| $nectar_options['header_format'] === 'centered-logo-between-menu' ) {
 			$usingPRCompatLayout = true;
 		}
 
-		if ( ! empty( $nectar_options['header_format'] ) && $nectar_options['header_format'] == 'centered-menu-bottom-bar' ) {
+		if ( ! empty( $nectar_options['header_format'] ) && $nectar_options['header_format'] === 'centered-menu-bottom-bar' ) {
 			$usingTopLeftRightCompatLayout = true;
 		}
 
@@ -39,7 +46,7 @@ if ( function_exists( 'register_nav_menus' ) ) {
 				$nectar_menu_arr = array(
 					'top_nav'            => 'Top Navigation Menu',
 					'top_nav_pull_right' => 'Top Navigation Menu Pull Right',
-					'secondary_nav'      => 'Secondary Navigation Menu <br /> <small>Will only display if applicable header layout is selected.</small>',
+					'secondary_nav'      => 'Secondary Navigation Menu',
 					'off_canvas_nav'     => 'Off Canvas Navigation Menu',
 				);
 
@@ -54,7 +61,7 @@ if ( function_exists( 'register_nav_menus' ) ) {
 			} else {
 				$nectar_menu_arr = array(
 					'top_nav'        => 'Top Navigation Menu',
-					'secondary_nav'  => 'Secondary Navigation Menu <br /> <small>Will only display if applicable header layout is selected.</small>',
+					'secondary_nav'  => 'Secondary Navigation Menu',
 					'off_canvas_nav' => 'Off Canvas Navigation Menu',
 				);
 			}
@@ -65,13 +72,21 @@ if ( function_exists( 'register_nav_menus' ) ) {
 				$nectar_menu_arr = array(
 					'top_nav'            => 'Top Navigation Menu',
 					'top_nav_pull_right' => 'Top Navigation Menu Pull Right',
-					'secondary_nav'      => 'Secondary Navigation Menu <br /> <small>Will only display if applicable header layout is selected.</small>',
+					'secondary_nav'      => 'Secondary Navigation Menu',
+				);
+
+			} elseif ( $usingTopLeftRightCompatLayout == true ) {
+
+				$nectar_menu_arr = array(
+					'top_nav'           => 'Top Navigation Menu',
+					'top_nav_pull_left' => 'Top Navigation Menu Pull Left',
+					'off_canvas_nav'    => 'Off Canvas Navigation Menu',
 				);
 
 			} else {
 				$nectar_menu_arr = array(
 					'top_nav'       => 'Top Navigation Menu',
-					'secondary_nav' => 'Secondary Navigation Menu <br /> <small>Will only display if applicable header layout is selected.</small>',
+					'secondary_nav' => 'Secondary Navigation Menu',
 				);
 			}
 		}
@@ -88,8 +103,11 @@ if ( function_exists( 'register_nav_menus' ) ) {
 
 
 
-
-// dropdown arrows
+/**
+ * Walker for adding in dropdown arrows.
+ *
+ * @since 5.0
+ */
 if ( ! function_exists( 'nectar_walker_nav_menu' ) ) {
 	function nectar_walker_nav_menu() {
 
@@ -102,11 +120,11 @@ if ( ! function_exists( 'nectar_walker_nav_menu' ) ) {
 						$header_format  = ( ! empty( $nectar_options['header_format'] ) ) ? $nectar_options['header_format'] : 'default';
 						$dropdownArrows = ( ! empty( $nectar_options['header-dropdown-arrows'] ) && $header_format != 'left-header' ) ? $nectar_options['header-dropdown-arrows'] : 'inherit';
 
-				if ( $header_format == 'centered-menu-bottom-bar' ) {
+				if ( $header_format === 'centered-menu-bottom-bar' ) {
 					$theme_skin = 'material';
 				}
 
-				if ( $theme_skin == 'material' ) {
+				if ( $theme_skin === 'material' ) {
 					$theme_skin = 'ascend';
 				}
 
@@ -119,16 +137,16 @@ if ( ! function_exists( 'nectar_walker_nav_menu' ) ) {
 				}
 
 				if ( ! empty( $children_elements[ $element->$id_field ] ) && $element->menu_item_parent == 0 && $theme_skin != 'ascend' && $header_format != 'left-header' && $dropdownArrows != 'dont_show' ||
-								! empty( $children_elements[ $element->$id_field ] ) && $element->menu_item_parent == 0 && $dropdownArrows == 'show' ) {
-					$element->title     = $element->title . '<span class="sf-sub-indicator"><i class="icon-angle-down"></i></span>';
+								! empty( $children_elements[ $element->$id_field ] ) && $element->menu_item_parent == 0 && $dropdownArrows === 'show' ) {
+					$element->title     = $element->title . '<span class="sf-sub-indicator"><i class="fa fa-angle-down icon-in-menu"></i></span>';
 					$element->classes[] = 'sf-with-ul';
 				}
 
 				if ( ! empty( $children_elements[ $element->$id_field ] ) && $element->menu_item_parent != 0 && $header_format != 'left-header' ) {
-					$element->title = $element->title . '<span class="sf-sub-indicator"><i class="icon-angle-right"></i></span>';
+					$element->title = $element->title . '<span class="sf-sub-indicator"><i class="fa fa-angle-right icon-in-menu"></i></span>';
 				}
 
-				if ( empty( $button_style ) && $header_format == 'left-header' ) {
+				if ( empty( $button_style ) && $header_format === 'left-header' ) {
 					$element->title = '<span>' . $element->title . '</span>';
 				}
 
@@ -145,14 +163,16 @@ nectar_walker_nav_menu();
 
 
 
-
-
-
-
+/**
+ * Add in description field into menu link output.
+ *
+ * @since 5.0
+ */
 if ( ! function_exists( 'nectar_description_walker_nav_menu' ) ) {
+	
 	function nectar_description_walker_nav_menu( $item_output, $item, $depth, $args ) {
-		if ( 'off_canvas_nav' == $args->theme_location && $item->description ) {
-			$item_output = str_replace( $args->link_after . '</a>', $args->link_after . '</a><small class="nav_desc">' . $item->description . '</small>', $item_output );
+		if ( 'off_canvas_nav' === $args->theme_location && $item->description ) {
+			$item_output = str_replace( $args->link_after . '</a>', $args->link_after . '</a><small class="nav_desc">' . wp_kses_post($item->description) . '</small>', $item_output );
 		}
 
 		return $item_output;
@@ -166,10 +186,11 @@ add_filter( 'walker_nav_menu_start_el', 'nectar_description_walker_nav_menu', 10
 
 
 
-
-
-
-
+/**
+ * Menu item style.
+ *
+ * @since 5.0
+ */
 if ( ! function_exists( 'nectar_nav_button_style' ) ) {
 
 	function nectar_nav_button_style( $output, $item, $depth, $args ) {
@@ -197,48 +218,57 @@ if ( ! function_exists( 'nectar_nav_button_style' ) ) {
 	}
 }
 
-	add_action( 'wp_nav_menu_item_custom_fields', 'nectar_nav_button_style', 10, 4 );
+add_action( 'wp_nav_menu_item_custom_fields', 'nectar_nav_button_style', 10, 4 );
 
 
 
 
 
 
-	$nectar_custom_menu_fields = array(
-		'menu-item-nectar-button-style' => '',
-	);
+$nectar_custom_menu_fields = array(
+	'menu-item-nectar-button-style' => '',
+);
 
-	function nectar_nav_button_style_update( $menu_id, $menu_item_db_id, $menu_item_args ) {
+/**
+ * Menu item style update.
+ *
+ * @since 5.0
+ */
+function nectar_nav_button_style_update( $menu_id, $menu_item_db_id, $menu_item_args ) {
+	
+	if( !function_exists('get_current_screen') ) {
+		return;
+	}
+	
+	$current_screen = get_current_screen();
 
-		$current_screen = get_current_screen();
+	// fix auto add new pages to top nav
+	$on_post_type = ( $current_screen && isset( $current_screen->post_type ) && ! empty( $current_screen->post_type ) ) ? true : false;
 
-		// fix auto add new pages to top nav
-		$on_post_type = ( $current_screen && isset( $current_screen->post_type ) && ! empty( $current_screen->post_type ) ) ? true : false;
+	global $nectar_custom_menu_fields;
 
-		global $nectar_custom_menu_fields;
+	if ( defined( 'DOING_AJAX' ) && DOING_AJAX || $on_post_type ) {
+		return;
+	}
+	check_admin_referer( 'update-nav_menu', 'update-nav-menu-nonce' );
 
-		if ( defined( 'DOING_AJAX' ) && DOING_AJAX || $on_post_type ) {
-			return;
+	foreach ( $nectar_custom_menu_fields as $key => $label ) {
+
+		// Sanitize
+		if ( ! empty( $_POST[ $key ][ $menu_item_db_id ] ) ) {
+			// Do some checks here...
+			$value = sanitize_text_field( $_POST[ $key ][ $menu_item_db_id ] );
+		} else {
+			$value = null;
 		}
-		check_admin_referer( 'update-nav_menu', 'update-nav-menu-nonce' );
 
-		foreach ( $nectar_custom_menu_fields as $key => $label ) {
-
-			// Sanitize
-			if ( ! empty( $_POST[ $key ][ $menu_item_db_id ] ) ) {
-				// Do some checks here...
-				$value = sanitize_text_field( $_POST[ $key ][ $menu_item_db_id ] );
-			} else {
-				$value = null;
-			}
-
-			// Update
-			if ( ! is_null( $value ) ) {
-				update_post_meta( $menu_item_db_id, $key, $value );
-			} else {
-				delete_post_meta( $menu_item_db_id, $key );
-			}
+		// Update
+		if ( ! is_null( $value ) ) {
+			update_post_meta( $menu_item_db_id, $key, $value );
+		} else {
+			delete_post_meta( $menu_item_db_id, $key );
 		}
 	}
+}
 
-	add_action( 'wp_update_nav_menu_item', 'nectar_nav_button_style_update', 10, 3 );
+add_action( 'wp_update_nav_menu_item', 'nectar_nav_button_style_update', 10, 3 );
